fix: update legend styles
