+++ packages/analytics/analytics-geo-map/src/composables/index.ts
@@ -1,6 +1,8 @@
 import useI18n from './useI18n'
+import useMetricUtils from './useMetricUtils'
 
 // All composables must be exported as part of the default object for Cypress test stubs
 export default {
   useI18n,
+  useMetricUtils,
 }
